aarch64: add _start entry point code
The linker uses _start as the default entry point for ELF binaries.
It is not a function and must invoke the exit system call at the end.
Its only task is to gather process parameters passed by the kernel,
executing the lone program and ensuring the process exits cleanly
should lone return a status code.