lone: define lone value structure
An extremely simple tagged union dynamic data type.
Bytes are the only type defined right now so it is alone in the union.
Not the most efficient representation but it will do for now.
Considered using tagged pointers but for now I just want it to work.
Maybe in the future.