lone: define the lone_lisp structure
This structure represents the programming language's internal state.

There is no dynamic memory allocation yet and even when implemented
it might not be available while the system is initializing.
So lone will need to be given memory to work with.
That means remembering the location and size of the memory block.