lone: add list-based lexical analysis function
Turns out I wasn't building a parser at all so let's make things right.

Input is text, a linear stream of characters.
The lexer accepts that input and returns a linear stream of tokens.
The parser builds nested structures out of that linear input.
So what I had was a lexer all along. A tokenizer.

Reworking this function gave me the chance to integrate lists.
The lexer accepts one bytes value and returns a list of bytes values.