lone: fix out of bounds array access bug
lexer->input.pointer[lexer->input.count] is not a valid index.
This was causing the lexer to read one extra character as a symbol
in the final iteration of the loop. In my tests it was a NUL character.
This explains the mysterious NULs that were showing up in output files.