lone: turn envp into a list of lone text values
This will give lone programs access to environment variables.

Should split them on '=' into key/value pairs in the future.
An association list or a hash table? Haven't decided yet.
Association lists are "pure" and won't complicate the lone code
but hash tables are just too useful and essential to be omitted.