lone: emit symbols in lexer instead of bytes
When the lexer finds a symbol, it has already parsed it.
There is no need for the lexer logic to be duplicated in the parser.
Just emit an actual symbol object into the tokens list.
When the parser sees that token, it knows it's already been parsed.

I didn't realize this before because when I was developing the lexer
I had only the bytes and list types defined. The system was not able to
distinguish different types of tokens yet. Now that I have several types
defined, the parser's work should be much simpler.

Now that I think about it, there should be no need for the parser to do
anything other than match parentheses and build up lists and trees.
All atoms should already be parsed into proper values by the lexer.
By the time they get to the parser, there should be no bytes objects
remaining in the tokens list unless I add support for a bytes literal.