lone: emit text objects directly to the token list
The lexer's byte consumption function is renamed to reflect the change.
The quote characters used to be included in the text in order to denote
the type of the token but that's no longer necessary. The parser doesn't
have to chop the quotes out of the string anymore.

All currently recognized atoms are already implemented in the lexer.
This enables me to remove three now superfluous parsing functions,
reducing the size of the code.