lone: exclude NUL when creating from C strings
In the lone system lengths are always known so there is no need to
include the trailing NUL byte which is the source of oh so many errors,
including the one I'm correcting right now.

Forget about the NUL byte. It should not end up in output either.
Maybe I'll silently append a hidden NUL at the end of to text buffers
once lone gets interoperability with C. For convenience only.