lone: define lone memory block structure
Lone needs to keep track of the memory it owns in order to manage it.
It will do so by maintaining a doubly linked list of metadata blocks.
These memory block descriptors are prefixed to their memory blocks and
allow lone to keep track of its size, location and availability for
allocation.