lone: define initialization function for lone lisp
Initializes the lone internal state by setting it up with the given
block of memory. Lone will take ownership of the given block and use it
to satisfy all memory allocation requests. In the future, lone will
probably gain the ability to map in more pages but for now this is the
only way it can dynamically create objects.