lone: keep track of all lone values allocated
Whenever a new lone value is created, add a link to it in lone's
internal state structure. This is accomplished by wrapping lone values
in a linked list header structure, linking and returning the address to
the actual lone value. This machinery is transparent to the caller who
receives real lone values.