lone: deallocate all objects at the end
Free all memory after the read-eval-print cycle finishes.
Lone doesn't have a proper REPL, environments or even a stack yet
so after the cycle pretty much all the objects are dead.

This also serves the useful purpose of verifying that the algorithms
are working as intended. At the end of the program, lone's memory blocks
should be completely empty and at maximum capacity.