lone: reallocate input buffers when full
The old code relied on implementation details of the previous memory
allocation algorithm to avoid the need for a reallocation function:
it assumed allocations would be contiguous. The previous sbrk-like
allocator had this property but now that lone has a proper allocator
this assumption is rendered invalid due to the presence of prefixed
memory block descriptors.

So now the read function must properly reallocate its buffers when full.