lone: don't subtract block descriptor size
The memory splitting function has already set up all sizes correctly.
This adjustment was causing small reductions to the size of blocks
and they were seriously adding up over time.

Lone now finishes with the same amount of memory it started with. Nice.
This branch is ready to be merged.