lone: define lone table set function
Makes sure the table is appropriately sized then sets the table entry
with the previously defined function. Maintains a 0.5 load factor.

The reason for the similar yet separate functions is the resizing code
also needs the table entries handling part of this logic. The returned
code allows it to know whether new data was added to the table so it can
update the count accordingly.