test: add automated testing bash script
The script automatically tests lone by executing the interpreter
and then comparing output and status code with expected values.
It's capable of supplying lone with arguments, environment and input
as well as capturing status code and standard output/error.
All input parameters and expected outputs are read from files
in each test's directory and they need not be present.

The structure of lone's test suite is as follows:

    test.sh                # the automated testing script
    test/                  # test suite
        x/y/z/             # test name, can be an arbitrary tree
            environment    # environment variables, one per line
            arguments      # arguments to pass, one per line
            input          # redirected to standard input
            output         # expected standard output
            error          # expected standard output
            status         # expected standard output

It's also capable of outputting test results in color
if standard output is connected to a terminal. Really nice.