lone: define the lone reader structure
Lone's gonna have to remember a lot more information if it is to have
a working REPL. So far the parser has been built assuming the whole
input is available. That isn't going to work for an interactive
programming language. It must be able to accept partial input
and ask for more. That means remembering partial parses and reading
until either a value or an error is produced.

Might as well take this opportunity to streamline the reader interface
by placing parameters in the structure as well as organize the sections.