lone: use the reader context in the read function
Read will now expect more input if the parser signals it's incomplete by
returning NULL. The loop reads input and tries to parse it. Success is
indicated by a valid lone lisp value being returned by the parser, in
that case the reader is reset and the value is returned. An incomplete
parse is signaled with NULL and in this case the reader will read more
input and try to parse it all from the beginning again. If there's no
more input available, it signals failure by also returning NULL.

Partial parser state is discarded. Unfortunately, I could not figure out
how to propagate this state in a way that allows the parser to restart.
Since I implemented the parser via recursive descent, the state and
mechanisms of the pushdown automaton, namely the functions and the call
stack, are not reified and cannot be saved for later continuation. Well,
not in C anyway. In any case, this is certainly a part of the language
that should be improved in the future.