lone: refactor parser from push to pull design
The previous architecture involved feeding the parser chunks of input
and getting back an arbitrary number of values as a partial results.
Keeping track of partial parse state and remaining input proved to be
a bothersome and complicated task.

The new design inverts the parser paradigm from push to pull.
Now the reader requests one value from the parser, which in turn
requests the lexer as many tokens as necessary to complete its task.
The lexer's peek/consume interface are the site where I/O actually
happens: when peeking would overrun the input buffer, the reader
transparently fills the buffer with more data by reading more input.

This design leads to the lone lisp reader naturally blocking when new
input is necessary to produce a value and the result is a functional
REPL without convoluted partial parser state tracking. Much simpler!

I took this opportunity to further simplify and reduce the size of the
code by merging the reader, parser and lexer into one section. They are
now more tightly integrated as they should have been from the start and
are easier to understand with less functions and less structures.