lone: evaluate the let special form
The let special form creates new environments, binds new variables and
evaluates a list of expressions in new contexts. Let accepts a list of
bindings in the form of symbols and values. Later bindings may refer to
earlier ones. All expressions after the bindings list are evaluated in a
new environment with those variables set. Let returns the value of the
last evaluated expression, or nil if none were evaluated.

Syntax:

    (let (x 10))                   ; nil
    (let (x 10) x);                ; 10
    (let (x 10 y 20) y)            ; 20
    (let (x 10 y x) y)             ; y, 10
    (let (x 10 y x z 20) x y z)    ; z, 20