lone: apply arguments to functions when evaluating
List forms are evaluated by the form evaluation function.
Special forms are special cases in the function, evaluated separately.
If it's not a special form though, it must be a function application.
The first element in the list is evaluated and the rest of the elements
are applied to the resulting function.

It is an error for the first element in a list to be a value other than
a special form symbol or a function.