test: ensure input test file is always included
Without the lone test program input file, the test will fail
because bash tries to redirect a non-existent file to standard input.
There is no point in running a test without a test program
that exercises the language as input anyway.

To accomplish this, the test parameters array is converted to an
associative array and the input parameter is always set unconditionally.