lone: define import special form evaluator
The import form searches the table of loaded modules for one matching
the specified symbol. If successful, it checks the remaining arguments.
If there are no arguments, it imports all the symbols in the module to
the current environment. If there are more symbols, it binds only those.

Binding means setting a symbol in the current environment to a value.

Syntax:

    (import lone)
    (import lone print)
    (import math)
    (import math + -)