lone: add system call table to lone lisp struct
The system call table will translate symbolic names
to their Linux system call numbers for the underlying platform.

Couldn't figure out a better place to put it
so I just threw it in the lone lisp structure for now.
I wanted to put it in the Linux module but I still haven't figured out
a good way to pass it to the system call primitive. I developed two now
abandoned branches trying to solve that problem but they led to nowhere.

Maybe I should turn primitives into a structure with a function pointer
and a lone table representing the function's closure.
Sort of like user data pointers.