lone: implement set as a primitive function
One less special case in the form evaluator!
Since it's now a variable in the environment,
it now enjoys the benefits of hash table lookups.
It's not being checked against every symbol anymore.

Set must now be imported from the lone module,
so all tests that use set must also be updated.