lone: define vector resizing function
This function allocates a new array of the specified capacity,
migrates all the elements in the vector to it, deallocates the vector's
array and replaces it with the new array.

Unset elements are initialized to zero.
If the new capacity is smaller than the vector's current length,
it is truncated: migration moves elements only up to the new length
and the references to all remaining ones are dropped.