lone: define ownership transfer function for bytes
The function simply takes ownership of the given pointer,
meaning it will be deallocated when the value is garbage collected.
Implement the normal creation function by copying the given buffer
and then transferring the copy to the new value.

This allows primitives that build new buffers to avoid copies
by letting them transfer the new buffer to the lone value.