lone: maintain hidden NUL at the end of buffers
The beauty of simple known size buffers just wasn't meant to be.
They fit like a glove with I/O system calls like read and write
but anything that deals with paths requires NUL-terminated C strings.

The easiest and simplest way to do this is to maintain a hidden NUL
at the end of all buffers. This allows pointers to be passed directly
to functions expecting a C string. The NUL won't be counted as part of
the buffer so it will remain completely invisible to lone code that
takes buffer lengths into account.

The concatenation function is one such function
so it must be changed to maintain the hidden NUL.

I actually thought I had seen the end of this NUL bullshit. Sigh.