lone: define module loading function
Lone's module loader takes a file descriptor as parameter,
reads lisp forms from it and evaluates them one by one
in the context of a module.

All files will start in the null module context.
I plan to add a module switching primitive that changes the context
the following expressions are evaluated in. Something like:

    ; files start in the null module
    (module example); activates the example module
    (set x 0); binds x in example's environment
    (module ()); back to the null module