lone: move compilation parameters to the top
This places them near the other preprocessor definitions such as the
FNV hashing algorithm parameters. Also make it so the macros are not
redefined if they already exist. This makes it possible to override
lone's memory and input buffer sizes without changing the source code
by methods such as specifying new definitions via compiler flags.
Should integrate this with the makefile later.