lone: define the module for name function
Returns the null module for null names.
Looks up the name in lone's loaded modules table and returns the module
if it's already been loaded. Otherwise, creates a new module, caches it
among lone's loaded modules and returns it to the caller.

Placing the modules in the loaded modules table ensures they are
reachable by the garbage collector. Returning existing modules also
implies they will only be created once and that they can be extended.