lone: load module by name for file descriptor
The file descriptor no longer implies a null module.
Lisp forms read from it can now be evaluated in any module's context.
It will now create and cache or reopen a module for the given name
and then evaluate the read expressions in its environment.
The null module will be used if name is null.