lone: unconditionally assign the next value
The garbage collector terminates when it reaches the last value
in lone's list of allocated values, marked by a null next pointer.
Ensure zero is assigned to the next pointer of the first created value.
It happened to work before because the memory was zero filled.
Marking newly allocated memory with some constant revealed the problem.