lone: define allocation function for lone heaps
Creates a new value heap which is a fixed size array of lone values.
Capacity can be expanded by linking new heaps to existing ones.
These will act as a pool of values from which lone will allocate.
These values are initialized as dead and unmarked. They are allocated
by marking them as live and returned them to the program.