lone: fix bug in heap deallocation logic
When the heap is dead, it is deallocated and unlinked and its next
pointer becomes the next heap. However, when I wrote this code,
I must have forgotten that the for loop has an epilogue of its own.
So after deallocation it sets up the next heap which may be null...
And then the for loop epilogue dereferences the heap, segfaulting lone.
This is more likely to happen with smaller heaps.

The redesigned loop fixes all tests and should fix the Lone CI too.