lone: load unknown modules when importing
The import primitive will now load modules that are unknown to the lone
interpreter. Modules are loaded by searching for their source files
in conventional file system locations, opening those files and
evaluating their contents in the environment of a new module object.

This marks the beginning of lone's modules and libraries system.