lone: add support for static data to bytes types
Add a bit indicating whether the value owns the memory it points to
so that the garbage collector can know whether it should deallocate it.
Update transfer and creation interfaces for the bytes types to indicate
whether the data being passed should be copied and managed by the value.
Update existing code to indicate that copies should be made since that
was the previous behavior. Any code dealing with C strings will now
automatically transfer the pointer without copies or deallocations.

Eliminates many copies and memory allocations and deallocations.
Greatly reduces memory usage by enabling values to simply pointing to
the static C strings without copying them to lone's memory.
Exploring lone's memory in the debugger is much easier now.