lone: improve the lone value hashing algorithm
The offset basis of the FNV-1a algorithm can serve as a partial hash:

    > the same hash results if X, Y, and Z are actually concatenated
    > and the FNV hash applied to the resulting string or if FNV is
    > calculated on an initial substring and the result used as the
    > offset_basis when calculating the FNV hash of the remainder

Make use of this fact to correctly hash all the lone value components
such as the data type and the actual data it contains, namely by running
them through the FNV-1a algorithm itself instead of using XOR directly,
thereby ensuring they are properly mixed with the FNV prime.

It seems the offset basis is itself derived from some arbitrary data
sourced from an email and that the actual bits don't matter.

    > FNV-0 is the same as FNV-1 but with offset_basis set to zero.

    > The offset_basis values for the n-bit FNV-1a algorithms
    > are computed by applying the n-bit FNV-0 algorithm
    > to the 32 octets representing the following character
    > string in ASCII [RFC20]:

    >     "chongo <Landon Curt Noll> /\\../\\"

    > In the general case, almost any offset_basis will serve
    > so long as it is non-zero. The choice of a non-standard
    > offset_basis may be beneficial in defending against some
    > attacks that try to induce hash collisions.

I should investigate the viability of hashing the random bytes from
the auxiliary vector and using the result as the initial offset basis.

References:

    https://datatracker.ietf.org/doc/draft-eastlake-fnv/