make: implement clang cross compilation support
The target is defined as the host machine by default.
However, if the target architecture is explicitly specified,
clang will be used to cross compile lone to that architecture.
In this case, the UAPI variable must also be defined and set to
the directory of the Linux user space API headers for the architecture.