lone: refactor import routines with export in mind
Now the list of symbols must be a subset of the module's exports list.
If no symbols are specified, import will now default to the module's
full exports list. It won't import every symbol from the module anymore.

The code is now shorter and simpler, has less special cases.
There's quadratic algorithm hiding in it though: it's scanning a vector
inside a loop. I plan to implement a set type for lone at some point.
Once that's done, I should replace the vector with a set.