make: restructure generated includes targets
Create an include directory in the build tree
and add it to the compiler's search path.
Output the generated files there.
Ensure the paths of all the dependencies are correct.
Update the lone source code to reflect these changes.
Delete the definitions which are now useless.

Now single generated file is inside the build tree!