tools: implement lone ELF segment creation tool
The purpose of this tool is to create an easily customizable segment
in the ELF executable that the Linux kernel will map into memory
when it loads the interpreter. A pointer to it can be easily obtained
via the program headers table pointer passed via the auxiliary vector.

Through this program, it should be possible to embed lone lisp code
into the interpreter itself and create self-contained applications.

This tool does the following:

 - Parses and validates an ELF executable
 - Finds its program headers table
 - Moves it to the end of the file
 - Appends two new entries to the table
 - Adjusts the ELF header's pointer to it
 - Adjusts the PHDR segment dimensions
 - Sets the first one to a PT_LOAD segment covering the table
 - Sets the second one to a custom lone lisp loadable segment
 - Outputs the result to a new file

This leaves an unused hole in the file where the original table was.
There's no getting around that without re-linking the entire executable.
Fortunately, mold has just added support for placeholder segments which
will efficiently and completely solve this problem without any waste.
This tool works though and should be useful if mold is unavailable.

Thanks: Rui Ueyama <ruiu@cs.stanford.edu>
GitHub-Issue: https://github.com/rui314/mold/issues/1148