tools: use PT_LOAD instead of PT_LONE
I was under the misconception that Linux mapped in the segments
described by all of the headers but it turns out it's just PT_LOAD:

https://github.com/torvalds/linux/blob/master/fs/binfmt_elf.c

    if (elf_ppnt->p_type != PT_LOAD)
        continue;

I guess that puts an end my dreams of a custom lone lisp ELF segment.
Now to find out why the interpreter is segfaulting without even running.