lone: implement embedded lone module loader
Walks the program header table, finds a lone segment,
reads a lone table that describes its contents,
looks up the run symbol on the table to obtain
a pair of integers which describe the address and size
of the code block that is to be executed.
That code is then loaded and evaluated as the run module.
The address is an offset from the end of the descriptor table.
The size is the number of bytes that make up the code.

If this data is embedded into the interpreter as a lone segment:

    { run (0 . 61) }
    (import
      (lone print))

    (print "hello from program header")

Running it will cause it to print "hello from program header"
before proceeding to load code from standard input.

It is now possible to embed code into the lone executable,
thereby creating of self-contained, standalone, freestanding,
redistributable Linux lisp applications that run on any Linux.