lone: stabilize lone table order
Instead of directly linearly probing a sparse array of
hash table entries, probe a sparse array of indices and
use them to index into a compact array of table entries.

This allows the compact entries array to be insertion ordered.
Efficient deletion is maintained: previous logic is applied to
the sparse indexes array while entries in the compact array
are simply shifted on top of the deleted entry.
Printing and garbage collection code is simplified.
This will also make iteration trivial to implement in the future.
Much better than a linked list!

This solution is a slightly modified version
of the solution described by Raymond Hettinger
in his 2012 message to the Python mailing list:

https://mail.python.org/pipermail/python-dev/2012-December/123028.html