lone/types: convert lone values into double words
Instead of folding values into pointers, they will now be two words:
a value word and a metadata word.

This approach is flexible, extensible, understood by debuggers
and easier to develop since there are less pointers involved.
It also allows use of a full register for immediate values instead
of limiting their range and supports further optimizations
such as compact homogeneous collections.

This comes at the cost of requiring two registers instead of one
but it seems performance is competitive with tagged pointers
if internet discussions are to be believed. For now, I believe
the benefits outweigh the costs.

References:

https://softwareengineering.stackexchange.com/q/448875
https://old.reddit.com/r/ProgrammingLanguages/comments/186nai0/tagged_pointers_vs_fat_pointers/