lone/value: update list operations and algorithms
Getting the list manipulation logic right proved
to be the most difficult part of this transition.

Lone values are not pointers anymore.
Therefore they are no longer nullable
and must be initialized with nil instead.
This created its own problems because of
my decision to make nil a register type:
nil was no longer a list with null pointers
but the language really needs it to be a list
because it wants to dereference those pointers.

I solved this in practice by simply special casing nil.
When the essential list functions get nil as an argument,
they pretend that it's a pair of two nil values instead.
Reader functions simply return the nil that's supposed
to be there without actually dereferencing any pointer.
Writer functions actually reify the nil into a null list
with the relevant element set.

Another consequence of that typing decision
is the fact nil is no longer considered a list.
It required special handling in type pedicates too.
The list functions now check the type of the argument
before dereferencing the list pointers.

Given these list primitives, the rest of the code should work well.
Many debugging hours were required to get to this point though.

Also updated is the list appender.
It keeps track of the list head
and efficiently extends its tail
while traversing it only once.
This change was required due to
the new non-pointer lone values.

The other functions are reimplemented
on top of the list appender function.