lone/value: define list destructuring function
Walks the list while assigning the values it contains to the values
pointed to by the arguments passed to the function. Also counts the
arguments as it goes to ensure that the list has neither less or more
values than it is expected to have.

This awesome function will dramatically simplify handling of
argument lists passed to C functions like primitive implementations.