lone/types: hash and compare symbols by identity
What's the point of deduplicating the symbols
if not to compare them by identity like integers?
Let's fix that. Symbols are now hashed and compared
by their pointers instead of byte-by-byte.

This uncovered a problem in the symbol interning routine:
it would create a new symbol to look it up in the table.
Naturally, the table would not be able to find it since
their addresses were different. So the symbol table
is now keyed by bytes objects instead. O(n) comparisons
will only happen when interning a symbol.