memory: implement dynamic memory layouts
Dynamic memory layout are simple state machines which work by
specializing the data storage strategy to the values written
in memory, significantly reducing memory usage and therefore
improving cache friendliness. For example, integers are stored
contiguously as arrays of fixed length native words.
Additional strategies may be added in the future.

This optimization is made possible by the fact most collections
tend to contain homogeneous data: all elements have the same type.
This assumption allows the implementation to concentrate the type
information in one place instead of storing it over and over again
in memory with the elements. The type metadata can be very easily
regenerated when the values are read out of the memory block.
This metadata centralization should cut memory usage in about half
since lone values are double words.

It deoptimizes gracefully in case this assumption is ever violated
at runtime: the memory layout dynamically changes to a heterogeneous
layout that supports arbitrary combinations of lone values.

As values are written to the memory block,
the implementation checks whether its type
matches the overall type of the collection.
If so, it maintains an optimized homogeneous
memory layout tailored for the type of the value.
If not, it replaces the memory block with
a new one comprising the complete value structures,
allowing the use of any combination of types.

References:

https://tratt.net/laurie/research/pubs/html/bolz_diekmann_tratt__storage_strategies_for_collections_in_dynamically_typed_languages/