make: explicitly depend on lone's entry point file
Fixes the segfaulting tests caused by previous changes
which resulted in the main lone.o file not being linked.

This simplifies the makefile, makes it more consistent.
Now the entire source/lone directory is a library
that can be easily linked to any number of entry points.
Tools and tests are just the entire lone interpreter
linked to a different main function.

They do say that every sufficiently complicated C program
is just a lisp interpreter in disguise...