test: add support for normalizing argv0 with env
After my suggestions in the coreutils mailing list,
the env utility gained the ability to control the
program name of the command being invoked, finally
giving it full control over the inputs of the new
process. Lone's ability to read the zeroth argument
can now be tested by this test suite.