types: add bytes memory read/write functions
Functions to reinterpret lone bytes memory
as integers of specific sizes and signedness
and either read or write the values at the
specified offset.

This really ought to be the simplest thing ever.
Let's hope the C compiler doesn't screw this up
with its "clever" optimizations. I've disabled
alias analysis so it should probably be OK.

Probably.