types: add bounds checking functions
Add functions that compute whether the given lone bytes value
contains the given address, offset or memory block.

A block of memory is defined as a specific number of bytes
starting at the given offset. For example, 8 bytes at 0x40
forms a memory block. This allows checking whether arbitrary
data types are contained by the lone bytes value. The result
of sizeof is the intended parameter.