scripts/test: find test executables in PATH
Use the shell's PATH variable to look up test executables.
Also implement inheritable executable specifier files.
If the executable for a test is not specified,
see if it's specified in the parent directories
all the way up to the root of the test suite.
This allows encoding the notion of a default
executable as a file in the test suite itself.