tools/lone-embed: fix starting values for limits
Looks like I got the order wrong when I first wrote this.

When calculating the start of the ELF file and its memory image,
the goal is to find the smallest valid addresses and offsets.
Therefore the limits should be initialized with the maximum value
for the integer type in order to ensure so that the min() function
will returns the correct result for any possible input ELF.

When calculating the end of the ELF file and its memory image,
the goal is to find the largest valid addresses and offsets.
Therefore the limits should be initialized with zero in order to
ensure that the max() function will returns the correct result
for any possible input ELF.