tools/lone-embed: move null segments check
The set_lone_segments function is the code which actually depends
on those null segments. That's where the check should take place.

The analyze function will be called multiple times in the future.
A null segment count of zero will also cause this tool to move
the program headers table to the end of the file in order to
add the required null segments, allowing it to proceed with
the patching operation instead of just failing. Therefore,
in the future this will be just a sanity check.

Also add a well-named function for this to make the code readable.