tools/lone-embed: fix starting values for limits
Allocate additional memory for two three extra program headers.
In case there aren't enough null segments in the input ELF,
the program header table will be moved to the end of the file
and two null segments will be added automatically.

The third segment is for a new PT_LOAD entry covering the table.