lone: reject nil input for syscall table lookup
An attempt to lookup the nil system call is probably a bug.
Don't treat it as zero.

Come to think of it, I don't remember if lone's nil
has a zero value when type punned to an integer.
And I'm the one who created the damn thing.
After reviewing the code, it turns out that
this was just totally wrong from the start.

Values are two words: one for metadata
and another for the immediate value.
Nils are just values with the nil
type set in the metadata word.
The value of the data word
is completely undefined.