lone: extract system structure from interpreter
When I started lone I tried to force myself to keep things as simple
as possible in order to maintain focus on the task at hand.
As a result, I wrote code that was tightly coupled on purpose because
I thought I'd get bogged down in library development if I didn't.
Or worse, build system development. Hey, it's certainly happened before.

I believe that approach was instrumental in conjuring up lone lisp
into the realm of existence. I have created a freestanding interpreter
that can successfully run programs. That's no small feat, no matter how
simple and slow it is. Now comes the truly difficult part: making it
useful for something other than running fibonacci calculations.
For that, I need to continuously improve the system, the interpreter,
the language, the modules and everything in-between.
It's truly a lifetime of work.

Before I start walking down that road, it is wise to finally pay back
some of the mounting tech debt in this project. I'll begin by separating
the lone lisp interpreter from the underlying lone system libraries.

Lone was initially conceived as a freestanding lisp interpreter.
As its complexity grew, a library of vital functions was implemented
in order to support basic program operations such as memory allocation
even in the absence of the standard C library. They were meant to be
tightly integrated with the interpreter but their utility increased
to the point they supported the development of tools such as lone-embed.

The utility of the lone system is likely to increase even further
as development of lone continues. My desire to use them to build
general Linux software is also likely to increase. Decoupling the
system functions from the interpreter is therefore desirable.

So I am reorganizing the project in order to make the lone runtime
usable without the lone interpreter. This commit marks the beginning
of the lone system, a library for freestanding Linux development.

The liblinux reborn, better than it ever was.

May this help pave the way for amazing Linux free software.