types: define C string literal to lone bytes macro
Constructs a compound initializer for the lone bytes structure
correctly initialized with the contents of the given C string.
Requires the size of the string to be known at compile time
so the argument must be an actual C string literal whose size
can be computed by the sizeof operator.