include/lone: simplify initializer macros
I tried to compose them with each other in order to keep it DRY
but it was just too broken due to the C prepocessor's text pasting
nature. Just do it in the simplest, most repetitive, most brute force
way imaginable by copy pasting the macros all over the place and
customizing each of them with the desired parameters.